Extracted AtomFeedLikeness to a public class,
since I'll soon need to use it from other test classes too.
