+++ AtomEventStore.UnitTests/AtomEventStore.UnitTests.csproj
@@ -85,6 +85,7 @@
     <Compile Include="AtomEventsInFilesTests.cs" />
     <Compile Include="AtomFeedBuilder.cs" />
     <Compile Include="AtomFeedComparer.cs" />
+    <Compile Include="AtomFeedLikeness.cs" />
     <Compile Include="AtomFeedTests.cs" />
     <Compile Include="AtomEventsInMemoryTests.cs" />
     <Compile Include="AtomLinkTests.cs" />
